requirements.txt
sqlalchemy
psycopg2-binary
selenium
alembic
python-dotenv
webdriver-manager
beautifulsoup4
undetected-chromedriver
pydantic